# Conflicts:
#	app/Http/Controllers/Api/BomController.php

+++ database/migrations/2024_10_28_095915_create_mo_components_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->id('mo_component_id');
             $table->unsignedBigInteger('mo_id');
             $table->foreign('mo_id')->references('mo_id')->on('manufacturing_orders')->onDelete('cascade');
-            $table->integer('to_consume');
+            $table->double('to_consume');
             $table->unsignedBigInteger('material_id');
             $table->foreign('material_id')->references('material_id')->on('materials')->onDelete('cascade');
-            $table->integer('reserved');
+            $table->double('reserved');
             $table->boolean('consumed');
             $table->timestamps();
         });
